Workaround for bug in serializing to JSON in Bridge.NET
* make all DTOs ObjectLiteral
* disable reflection

+++ src/SIL.Machine.WebApi.Server/Models/Build.cs
@@ -17,7 +17,7 @@
 		}
 
 		public string Id { get; set; }
-		public long Revision { get; set; }
+		public int Revision { get; set; }
 		public string EngineId { get; set; }
 		public int StepCount { get; set; }
 		public int CurrentStep { get; set; }
