+++ src/SIL.Machine.WebApi.Server/Models/IEntity.cs
@@ -5,6 +5,6 @@ namespace SIL.Machine.WebApi.Server.Models
 	public interface IEntity<out T> : ICloneable<T> where T : IEntity<T>
 	{
 		string Id { get; set; }
-		long Revision { get; set; }
+		int Revision { get; set; }
 	}
 }
